requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==4.7.2
scikit-learn
dill
pyaml
##-e .

